+++ src/js/game/getWinners.js
