+++ src/js/game/isEffectiveScore.js
@@ -1,4 +1,4 @@
-import { GAME } from './constant.js';
+import { GAME } from '../utils/constant.js';
 
 export const isEffectiveScore = (num) => {
   return num >= GAME.EFFECTIVE_SCORE;
