+++ src/js/game/printGameResult.js
@@ -1,19 +1,5 @@
-import { getWinners } from '../utils/getWinners.js';
-import { toggleVisibility } from '../utils/toggleVisibility.js';
-
-const restartGame = () => {
-  const $carNameInput = document.querySelector('#car-name-input');
-  const $racingCountInput = document.querySelector('#racing-count-input');
-  const $gameProcessScreen = document.querySelector('#game-process-screen');
-
-  toggleVisibility('racingCountSection');
-  toggleVisibility('gameProcessSection');
-  toggleVisibility('gameResultSection');
-
-  $carNameInput.value = '';
-  $racingCountInput.value = '';
-  $gameProcessScreen.innerHTML = '';
-};
+import { getWinners } from './getWinners.js';
+import { restartGame } from './restartGame.js';
 
 export const printGameResult = () => {
   const $gameResultText = document.querySelector('#game-result-text');
