+++ src/js/utils/toggleVisibility.js
@@ -1,8 +1,8 @@
-const $racingCountSection = document.querySelector('#racing-count-section');
-const $gameProcessSection = document.querySelector('#game-process-section');
-const $gameResultSection = document.querySelector('#game-result-section');
-
 export const toggleVisibility = (target) => {
+  const $racingCountSection = document.querySelector('#racing-count-section');
+  const $gameProcessSection = document.querySelector('#game-process-section');
+  const $gameResultSection = document.querySelector('#game-result-section');
+
   const sectionList = {
     racingCountSection: $racingCountSection,
     gameProcessSection: $gameProcessSection,
